#0.5.0: restructured code
