+++ lib/fluent_assertions.dart
@@ -3,6 +3,6 @@
 /// It uses Dart's Extension Functions to provide a fluent wrapper around test assertions.
 library fluent_assertions;
 
-export 'src/basic_assertions.dart';
-export 'src/numerical_assertions.dart';
-export 'src/string_assertions.dart';
+export 'src/base/base_assertions.dart';
+export 'src/numerical/numerical_assertions.dart';
+export 'src/string/string_assertions.dart';
